+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.0.0] - Unreleased
 ### Added
 - Add methods that generate an Astarte Device ID using either a random UUIDv4 or by deriving an UUIDv5 using unique data.
+- Astarte Device register.
 
 ## [1.0.0-rc.0] - 2021-05-10
 ### Added